refactor : ChessBoardElement 패키지 변경 및 예외 메시지 추가
